Rename the directory.  Also remove test for Sweeping class since user will not use that directly.  Now user will use run_sweep within QAnalysis.  QAnalysis will make an instance of Sweeper class now.
