+++ qiskit_metal/analyses/__init__.py
@@ -98,4 +98,4 @@ from .hamiltonian.transmon_charge_basis import Hcpb
 from .hamiltonian import HO_wavefunctions
 from .hamiltonian import transmon_analytics
 from .hamiltonian.transmon_CPB_analytic import Hcpb_analytic
-from .sweep_options.sweeping import Sweeping
+from .sweep_and_optimize.sweeper import Sweeper
